# server/internal/schema/openapi.yaml
...
openapi: 3.0.0
info:
  title: Railway Wiki API
  description: 'Railway wiki API. Public **view** endpoints under `/api/{resource}` are read-only and
    open; **management** endpoints under `/api/management/{resource}` require a bearer token whose
    principal has the `admin` role and provide full CRUD.


    List/search responses use cursor pagination: `{ items, pagination: { next } }`. Errors use `{ error,
    code }`. Single-object responses are returned bare (never wrapped in a `data` envelope).'
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local development server
...
security:
- bearerAuth: []
paths:
...
  /api/management/media/upload-url:
    post:
      summary: Create a presigned upload URL for a media file
      description: Returns a short-lived presigned URL the client uses to upload the file bytes directly
        to object storage. After a successful upload, create the Media record with `adminCreateMedia`
        using the returned `url`.
      operationId: adminCreateMediaUploadUrl
      tags:
      - Media (Management)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignMediaUploadRequest'
      responses:
        '201':
          description: Presigned upload target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadTarget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer authentication; management endpoints require role `admin`.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required:
      - error
      - code
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          example: not_found
...
    PresignMediaUploadRequest:
      type: object
      required:
      - fileName
      - contentType
      properties:
        fileName:
          type: string
          description: Original file name, used to derive the object key.
          example: station-photo.jpg
        contentType:
          type: string
          description: MIME type of the file to be uploaded.
          example: image/jpeg
        sizeBytes:
          type: integer
          format: int64
          description: Expected file size in bytes, if known.
        mediaType:
          type: string
          description: Intended media type (e.g. image, video).
    MediaUploadTarget:
      type: object
      required:
      - uploadUrl
      - method
      - url
      - expiresAt
      properties:
        uploadUrl:
          type: string
          format: uri
          description: Presigned URL to upload the file bytes to.
        method:
          type: string
          description: HTTP method to use for the upload.
          example: PUT
        headers:
          type: object
          additionalProperties:
            type: string
          description: Headers that must be sent verbatim with the upload request.
        url:
          type: string
          format: uri
          description: Canonical URL of the object once uploaded; store this as Media.url.
        key:
          type: string
          description: Storage object key/path.
        expiresAt:
          type: string
          format: date-time
          description: When the presigned URL expires.