# server/scripts/gen_handlers.py
#!/usr/bin/env python3
"""Generate internal/server/server.go: Server struct, constructor, and all
resource handler methods for railway-wiki."""

# resource: (Singular, Plural, field, searchCols, filters[(ParamField, column)])
RES = [
    ("Company", "Companies", "companies", ["name", "short_name"], []),
    ("Station", "Stations", "stations", ["name", "name_en"], []),
    ("StationCode", "StationCodes", "stationCodes", [], [("StationId", "station_id"), ("RouteId", "route_id"), ("CompanyId", "company_id")]),
    ("StationTransfer", "StationTransfers", "stationTransfers", [], [("FromStationId", "from_station_id"), ("ToStationId", "to_station_id")]),
    ("Platform", "Platforms", "platforms", [], [("StationId", "station_id")]),
    ("Route", "Routes", "routes", ["name", "name_en"], []),
    ("RouteCompany", "RouteCompanies", "routeCompanies", [], [("RouteId", "route_id"), ("CompanyId", "company_id")]),
    ("RouteStation", "RouteStations", "routeStations", [], [("RouteId", "route_id"), ("StationId", "station_id")]),
    ("TrackSegment", "TrackSegments", "trackSegments", [], [("RouteId", "route_id"), ("FromStationId", "from_station_id"), ("ToStationId", "to_station_id")]),
    ("PlatformTrack", "PlatformTracks", "platformTracks", [], [("PlatformId", "platform_id"), ("TrackSegmentId", "track_segment_id")]),
    ("OperationRoute", "OperationRoutes", "operationRoutes", ["name", "name_en"], []),
    ("OperationRouteCompany", "OperationRouteCompanies", "operationRouteCompanies", [], [("OperationRouteId", "operation_route_id"), ("CompanyId", "company_id")]),
    ("OperationRouteSection", "OperationRouteSections", "operationRouteSections", [], [("OperationRouteId", "operation_route_id"), ("RouteId", "route_id")]),
    ("OperationRouteStop", "OperationRouteStops", "operationRouteStops", [], [("OperationRouteId", "operation_route_id"), ("StationId", "station_id")]),
    ("TimetableVersion", "TimetableVersions", "timetableVersions", ["name"], [("CompanyId", "company_id")]),
    ("ServiceCalendar", "ServiceCalendars", "serviceCalendars", ["name"], []),
    ("ServiceCalendarException", "ServiceCalendarExceptions", "serviceCalendarExceptions", [], [("CalendarId", "calendar_id")]),
    ("Train", "Trains", "trains", ["train_number", "train_name"], []),
    ("TrainRun", "TrainRuns", "trainRuns", [], [("TimetableVersionId", "timetable_version_id"), ("OperationRouteId", "operation_route_id"), ("TrainId", "train_id"), ("CalendarId", "calendar_id")]),
    ("TrainRunStop", "TrainRunStops", "trainRunStops", [], [("TrainRunId", "train_run_id"), ("StationId", "station_id")]),
    ("Media", "Media", "media", ["title"], [("MediaType", "media_type")]),
    ("MediaAttachment", "MediaAttachments", "mediaAttachments", [], [("MediaId", "media_id"), ("EntityType", "entity_type"), ("EntityId", "entity_id")]),
]

HAS_Q = {"Company", "Station", "Route", "OperationRoute", "TimetableVersion", "ServiceCalendar", "Train", "Media"}


def filter_expr(sing, filters):
    parts = []
    if filters:
        eq = ", ".join(f'"{col}": r.Params.{pf}' for pf, col in filters)
        parts.append(f"Eq: map[string]any{{{eq}}}")
    if sing in HAS_Q:
        parts.append("Q: r.Params.Q")
    if sing == "Station":
        parts.extend(["South: r.Params.South", "West: r.Params.West", "North: r.Params.North", "East: r.Params.East"])
    parts.append("Cursor: r.Params.Cursor")
    parts.append("Limit: r.Params.Limit")
    return "repo.Filter{" + ", ".join(parts) + "}"


def list_body(prefix, plural, field, fexpr):
    return f"""func (s *Server) {prefix}{plural}(ctx context.Context, r api.{prefix}{plural}RequestObject) (api.{prefix}{plural}ResponseObject, error) {{
\titems, next, err := s.{field}.List(ctx, {fexpr})
\tif err != nil {{
\t\treturn nil, err
\t}}
\treturn api.{prefix}{plural}200JSONResponse{{Items: items, Pagination: api.Pagination{{Next: next}}}}, nil
}}
"""


def get_body(prefix, sing, field):
    return f"""func (s *Server) {prefix}{sing}(ctx context.Context, r api.{prefix}{sing}RequestObject) (api.{prefix}{sing}ResponseObject, error) {{
\tm, err := s.{field}.Get(ctx, r.Id.String())
\tif err != nil {{
\t\treturn nil, err
\t}}
\treturn api.{prefix}{sing}200JSONResponse(*m), nil
}}
"""


out = []
out.append('''// Package server implements api.StrictServerInterface.
//
// Handlers are intentionally thin: they translate the generated request object
// into a repo.Filter or an id, call the generic service, and wrap the result in
// the operation's typed success response. All error-to-HTTP-status mapping is
// centralized in the ErrorMiddleware (errors.go); handlers just return an error.
//
// This file is generated by scratchpad/gen_handlers.py. Regenerate rather than
// hand-editing the repetitive resource handlers.
package server

import (
\t"context"

\t"github.com/rxtech-lab/railway-wiki/internal/api"
\t"github.com/rxtech-lab/railway-wiki/internal/convert"
\t"github.com/rxtech-lab/railway-wiki/internal/media"
\t"github.com/rxtech-lab/railway-wiki/internal/models"
\t"github.com/rxtech-lab/railway-wiki/internal/overpass"
\t"github.com/rxtech-lab/railway-wiki/internal/repo"
\t"github.com/rxtech-lab/railway-wiki/internal/schema"
\t"github.com/rxtech-lab/railway-wiki/internal/service"
\t"gorm.io/gorm"
)

// Ensure Server implements StrictServerInterface.
var _ api.StrictServerInterface = (*Server)(nil)
''')

# type aliases
out.append("// Per-resource service type aliases.")
for sing, plural, field, cols, filters in RES:
    out.append(f"type svc{sing} = *service.Service[models.{sing}, *models.{sing}, api.{sing}, api.Create{sing}Request, api.Update{sing}Request]")
out.append("")

# struct
out.append("// Server holds every resource service plus infra dependencies.")
out.append("type Server struct {")
out.append("\tdb *gorm.DB")
out.append("\tschemas *schema.Registry")
out.append("\tpresign media.Presigner")
out.append("\toverpass overpass.Client")
for sing, plural, field, cols, filters in RES:
    out.append(f"\t{field} svc{sing}")
out.append("}")
out.append("")

# constructor
out.append("// NewServer builds the Server, constructing all resource services from db.")
out.append("func NewServer(db *gorm.DB, presign media.Presigner, schemas *schema.Registry, overpassClient overpass.Client) *Server {")
out.append("\treturn &Server{")
out.append("\t\tdb: db,")
out.append("\t\tschemas: schemas,")
out.append("\t\tpresign: presign,")
out.append("\t\toverpass: overpassClient,")
for sing, plural, field, cols, filters in RES:
    colargs = "".join(f', "{c}"' for c in cols)
    out.append(f"\t\t{field}: service.New(repo.New[models.{sing}](db{colargs}), convert.{sing}ToAPI, convert.{sing}FromCreate, convert.{sing}ApplyUpdate),")
out.append("\t}")
out.append("}")
out.append("")

# health
out.append('''// GetHealth reports service health.
func (s *Server) GetHealth(ctx context.Context, r api.GetHealthRequestObject) (api.GetHealthResponseObject, error) {
\treturn api.GetHealth200JSONResponse{Status: "ok"}, nil
}
''')

# presign
out.append('''// AdminCreateMediaUploadUrl returns a presigned upload target for a media file.
func (s *Server) AdminCreateMediaUploadUrl(ctx context.Context, r api.AdminCreateMediaUploadUrlRequestObject) (api.AdminCreateMediaUploadUrlResponseObject, error) {
\tif r.Body == nil {
\t\treturn nil, repo.ErrBadRequest
\t}
\ttarget, err := s.presign.Presign(ctx, media.PresignInput{
\t\tFileName:    r.Body.FileName,
\t\tContentType: r.Body.ContentType,
\t\tMediaType:   r.Body.MediaType,
\t\tSizeBytes:   r.Body.SizeBytes,
\t})
\tif err != nil {
\t\treturn nil, err
\t}
\treturn api.AdminCreateMediaUploadUrl201JSONResponse(*target), nil
}
''')

# per-resource handlers
for sing, plural, field, cols, filters in RES:
    fexpr = filter_expr(sing, filters)
    out.append(f"// --- {sing} ---\n")
    out.append(list_body("List", plural, field, fexpr))
    out.append(get_body("Get", sing, field))
    out.append(list_body("AdminList", plural, field, fexpr))
    out.append(get_body("AdminGet", sing, field))
    out.append(f"""func (s *Server) AdminCreate{sing}(ctx context.Context, r api.AdminCreate{sing}RequestObject) (api.AdminCreate{sing}ResponseObject, error) {{
\tif r.Body == nil {{
\t\treturn nil, repo.ErrBadRequest
\t}}
\tm, err := s.{field}.Create(ctx, r.Body)
\tif err != nil {{
\t\treturn nil, err
\t}}
\treturn api.AdminCreate{sing}201JSONResponse(*m), nil
}}
""")
    out.append(f"""func (s *Server) AdminUpdate{sing}(ctx context.Context, r api.AdminUpdate{sing}RequestObject) (api.AdminUpdate{sing}ResponseObject, error) {{
\tif r.Body == nil {{
\t\treturn nil, repo.ErrBadRequest
\t}}
\tm, err := s.{field}.Update(ctx, r.Id.String(), r.Body)
\tif err != nil {{
\t\treturn nil, err
\t}}
\treturn api.AdminUpdate{sing}200JSONResponse(*m), nil
}}
""")
    out.append(f"""func (s *Server) AdminDelete{sing}(ctx context.Context, r api.AdminDelete{sing}RequestObject) (api.AdminDelete{sing}ResponseObject, error) {{
\tif err := s.{field}.Delete(ctx, r.Id.String()); err != nil {{
\t\treturn nil, err
\t}}
\treturn api.AdminDelete{sing}204Response{{}}, nil
}}
""")
    out.append(f"""func (s *Server) AdminGet{sing}Schema(ctx context.Context, r api.AdminGet{sing}SchemaRequestObject) (api.AdminGet{sing}SchemaResponseObject, error) {{
\treturn api.AdminGet{sing}Schema200JSONResponse(s.schemas.For("{sing}", string(r.Params.Action))), nil
}}
""")

with open("internal/server/server.go", "w") as f:
    f.write("\n".join(out))
print("wrote internal/server/server.go")